+++ catalyst-gateway/bin/src/metrics/chain_follower.rs
@@ -1,1 +1,1 @@
-
+//! Metrics related to Chain Follower analytics.
