+++ catalyst-gateway/bin/src/metrics/chain_indexer.rs
@@ -1,1 +1,1 @@
-
+//! Metrics related to Chain Indexer analytics.
