+++ catalyst-gateway/bin/src/metrics/memory.rs
@@ -1,1 +1,1 @@
-
+//! Metrics related to memory analytics.
