+++ catalyst-gateway/bin/src/metrics/mod.rs
@@ -1,6 +1,18 @@
-//! This module contains submodules related to metrics report.
+//! This module contains submodules related to metrics report and analytics.
+
+use prometheus::{default_registry, Registry};
 
 pub(crate) mod chain_follower;
 pub(crate) mod chain_indexer;
 pub(crate) mod endpoint;
 pub(crate) mod memory;
+
+/// Initialize Prometheus metrics.
+///
+/// ## Returns
+///
+/// Returns the default prometheus registry.
+#[must_use]
+pub(crate) fn init_prometheus() -> Registry {
+    default_registry().clone()
+}
